requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.9
django-debug-toolbar==3.2.2
djangorestframework==3.12.4
# djangorestframework-simplejwt==5.0.0
djangorestframework_simplejwt==5.0.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.2
MarkupSafe==2.0.1
openpyxl==3.0.9
packaging==21.2
prompt-toolkit==3.0.22
psycopg2-binary==2.9.2
PyJWT==2.3.0
pyparsing==2.4.7
python-dotenv==0.19.2
pytz==2021.3
q==2.6
redis==3.5.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
